% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{add_boot_ci}
\alias{add_boot_ci}
\title{Adds bootstrap confidence intervals.}
\usage{
add_boot_ci(grouped_data, funcs)
}
\description{
Adds bootstrap confidence intervals.
}
\keyword{internal}
