% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{add_bootstrap_data}
\alias{add_bootstrap_data}
\title{Adds bootstrap data to the model that will be used for plotting.}
\usage{
add_bootstrap_data(neural_net, options)
}
\description{
Adds bootstrap data to the model that will be used for plotting.
}
\keyword{internal}
