% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R,
%   R/neural_net_representative_values.R
\name{change}
\alias{change}
\title{Change input data via change_variables and change_values}
\usage{
change(data, change_variables, change_values, class, new_data, rep, type)

change(data, change_variables, change_values, class, new_data, rep, type)
}
\description{
Change input data via change_variables and change_values

change input data via change_variables and change_values
}
\keyword{internal}
