% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R,
%   R/neural_net_representative_values.R
\name{colMedians}
\alias{colMedians}
\title{Columnwise medians}
\usage{
colMedians(x)

colMedians(x)
}
\description{
Columnwise medians

columnwise medians
}
\keyword{internal}
