% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R,
%   R/neural_net_representative_values.R
\name{colModes}
\alias{colModes}
\title{Columnwise modes}
\usage{
colModes(x)

colModes(x)
}
\description{
Columnwise modes

columnwise modes
}
\keyword{internal}
