% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{compute_boot_ci}
\alias{compute_boot_ci}
\title{Computes bootstrap interval for given probability.}
\usage{
compute_boot_ci(bootstrap_data, probs, nrepititions)
}
\description{
Computes bootstrap interval for given probability.
}
\keyword{internal}
