% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{compute_bootstrap_ci}
\alias{compute_bootstrap_ci}
\title{computes Bootstrap intervals}
\usage{
compute_bootstrap_ci(partial_dependence, grid, predictor, neural_net,
  probs, nrepetitions)
}
\description{
computes Bootstrap intervals
}
\keyword{internal}
