% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{compute_bootstrap_ci_repres}
\alias{compute_bootstrap_ci_repres}
\title{create bootstrap samples}
\usage{
compute_bootstrap_ci_repres(result, predictor, predictor_value, neural_net,
  kind, rep, units, change_values, change_variables, class, probs,
  nrepetitions)
}
\description{
create bootstrap samples
}
\keyword{internal}
