% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{compute_mean_prediction_numeric}
\alias{compute_mean_prediction_numeric}
\title{Computes mean prediction for numeric response}
\usage{
compute_mean_prediction_numeric(grid, neural_net, predictor)
}
\description{
Computes mean prediction for numeric response
}
\keyword{internal}
