% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{create_functions}
\alias{create_functions}
\title{Create functions used for calculating relevant metrics.}
\usage{
create_functions(probs, nrepititions)
}
\description{
Create functions used for calculating relevant metrics.
}
\keyword{internal}
