% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{create_grid}
\alias{create_grid}
\title{Creates the grid used for predicting yhat.}
\usage{
create_grid(neural_net, predictor)
}
\description{
Creates the grid used for predicting yhat.
}
\keyword{internal}
