% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{create_plotting_data}
\alias{create_plotting_data}
\title{Creates the plotting data.}
\usage{
create_plotting_data(grid, predictor, neural_net, probs, nrepetitions)
}
\description{
Creates the plotting data.
}
\keyword{internal}
