% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{descale}
\alias{descale}
\title{Descales selected predictor column.}
\usage{
descale(predictor, neural_net, grid, prediction_scale = F)
}
\description{
Descales selected predictor column.
}
\keyword{internal}
