% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{descale_grid}
\alias{descale_grid}
\title{Descales selected predictor column and prediction value for grid.}
\usage{
descale_grid(predictor, neural_net, grid)
}
\description{
Descales selected predictor column and prediction value for grid.
}
\keyword{internal}
