% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{empty_predictor_value}
\alias{empty_predictor_value}
\title{warns about length of predictors for plotting}
\usage{
empty_predictor_value(neural_net, predictor, class)
}
\description{
warns about length of predictors for plotting
}
\keyword{internal}
