% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworkVisualization.R
\name{example_nn_model}
\alias{example_nn_model}
\title{Returns example models by type of response}
\usage{
example_nn_model(type)
}
\arguments{
\item{type}{String beeing either numerical, categorical or binary.}
}
\value{
NeuralNetwork class
}
\description{
Each model has stored data available with 1000 bootstrap iterations and a
90\% confidence interval. The code that fitted the models can be found here:
\url{https://github.com/AlexAfanasev/NeuralNetworkVisualization/blob/master/inst/examples.R}
}
\examples{
\dontrun{
# Example: Numeric
example_nn_model("numerical")

# Example: Categoric
example_nn_model("categorical")

Example: Binary
example_nn_model("binary")

}

}
