% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{fit_neural_network}
\alias{fit_neural_network}
\title{Fits neural network for either numerical or categorical dependent variable.}
\usage{
fit_neural_network(f, data, layers, type, dependent, independent, ...)
}
\description{
Fits neural network for either numerical or categorical dependent variable.
}
\keyword{internal}
