% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{fit_neural_network_categorical}
\alias{fit_neural_network_categorical}
\title{Fits neural network for cateogircal dependent variable.}
\usage{
fit_neural_network_categorical(f, data, layers, dependent, independent,
  ...)
}
\description{
Fits neural network for cateogircal dependent variable.
}
\keyword{internal}
