% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{fit_neural_network_numeric}
\alias{fit_neural_network_numeric}
\title{Fits neural network for numerical dependent variable.}
\usage{
fit_neural_network_numeric(f, data, layers, ...)
}
\description{
Fits neural network for numerical dependent variable.
}
\keyword{internal}
