% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{get_independent}
\alias{get_independent}
\title{Returns the independent variables based on specification.}
\usage{
get_independent(data, dependent_variable, specification)
}
\description{
Returns the independent variables based on specification.
}
\keyword{internal}
