% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{get_predictors}
\alias{get_predictors}
\title{Returns the predictors for which to plot the partial dependencies.}
\usage{
get_predictors(neural_net, predictors)
}
\description{
Returns the predictors for which to plot the partial dependencies.
}
\keyword{internal}
