% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R
\name{grid_categorical}
\alias{grid_categorical}
\title{customize grid for categorical variables}
\usage{
grid_categorical(class, neural_input, cols_numeric, kind)
}
\description{
customize grid for categorical variables
}
\keyword{internal}
