% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R
\name{grid_factor}
\alias{grid_factor}
\title{customize grid for dummy variables}
\usage{
grid_factor(neural_net, mean_input, class, type, rep)
}
\description{
customize grid for dummy variables
}
\keyword{internal}
