% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{is_valid_nrepetitions}
\alias{is_valid_nrepetitions}
\title{Checks for valid nrepetitions for the bootstrap confidence interval.}
\usage{
is_valid_nrepetitions(nrepetitions)
}
\description{
Checks for valid nrepetitions for the bootstrap confidence interval.
}
\keyword{internal}
