% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_neural_net_visualization.R
\name{is_valid_probs}
\alias{is_valid_probs}
\title{Checks for valid prediction interval probabilities.}
\usage{
is_valid_probs(probs)
}
\description{
Checks for valid prediction interval probabilities.
}
\keyword{internal}
