% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R,
%   R/neural_net_representative_values.R
\name{make_grid}
\alias{make_grid}
\title{Create input grid}
\usage{
make_grid(neural_net, kind, class = NULL, type)

make_grid(neural_net, kind, class = NULL, type)
}
\description{
Create input grid

create input grid
}
\keyword{internal}
