% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{plan_process}
\alias{plan_process}
\title{Function for enabling multiprocessing.}
\usage{
plan_process(parallel)
}
\description{
Function for enabling multiprocessing.
}
\keyword{internal}
