% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{plot.NeuralNetwork}
\alias{plot.NeuralNetwork}
\title{Plot method for neural network.}
\usage{
\method{plot}{NeuralNetwork}(x, ...)
}
\arguments{
\item{x}{Fitted NeuralNetwork model}

\item{...}{further parameters for plot neuralnet, see:
\code{\link[neuralnet]{plot.nn}}}
}
\description{
Plot method for neural network.
}
