% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{plot_multiple_categorical}
\alias{plot_multiple_categorical}
\title{Plots partial dependencies for given predictor with categorical dependent
variable.}
\usage{
plot_multiple_categorical(prepared_data, neural_net)
}
\description{
Plots partial dependencies for given predictor with categorical dependent
variable.
}
\keyword{internal}
