% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{plot_multiple_numerical}
\alias{plot_multiple_numerical}
\title{Plots partial dependencies for given predictors with numerical dependent
variable.}
\usage{
plot_multiple_numerical(prepared_data, neural_net)
}
\description{
Plots partial dependencies for given predictors with numerical dependent
variable.
}
\keyword{internal}
