% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{plot_multiple_res}
\alias{plot_multiple_res}
\title{create plotting for multiple predictors}
\usage{
plot_multiple_res(neural_net, predictor, predictor_value, rep, units, kind,
  change_variables, change_values, class, probs, nrepetitions)
}
\description{
create plotting for multiple predictors
}
\keyword{internal}
