% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{plot_partial_dependencies}
\alias{plot_partial_dependencies}
\title{Function for plotting the partial dependencies of the created NeuralNetwork}
\usage{
plot_partial_dependencies(neural_net, predictors = "all", probs = c(0,
  0), type = "ggplot", nrepetitions = 50)
}
\arguments{
\item{neural_net}{The NeuralNetwork instance, see:
\code{\link{NeuralNetwork}}}

\item{predictors}{Vector of predictors of the neural network for which to
plot the partial dependencies.}

\item{probs}{Vector of lower and upper bound probabilities for the confidence
interval. If booth are 0, intervals will not be plotted.}

\item{type}{Either 'ggplot' if the plot should be created using ggplot or
'ggplotly' if plotly should be used.}

\item{nrepetitions}{Number of samples used within bootstrap for confidence
intervals.}
}
\value{
Created figure
}
\description{
\code{plot_partial_dependencies} plots the partial dependencies for
  the specified predictors.
}
\examples{
\dontrun{
# Example: Numeric or Binary
library(MASS)
neural_network <- NeuralNetwork(f = "medv ~ .", data = Boston,
                                layers = c(5, 3), scale = TRUE,
                                linear.output = TRUE)
plot_partial_dependencies(neural_network, predictors = "crim",
                          probs = c(0.2, 0.8), type = "ggplotly")
plot_partial_dependencies(neural_network, predictors = c("crim", "age"))
plot_partial_dependencies(neural_network, probs = c(0.1, 0.9))

# Example: Categoric
library(datasets)
model <- NeuralNetwork(
   Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
   data = iris, layers = c(10, 10), rep = 5, err.fct = "ce",
   linear.output = FALSE, lifesign = "minimal", stepmax = 1000000,
   threshold = 0.001, scale = F)

plot_partial_dependencies(model, predictors = "Petal.Length")
plot_partial_dependencies(model,
                          predictors = c("Sepal.Length", "Petal.Length"))
plot_partial_dependencies(model, type = "ggplotly")
}

}
