% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot_partial_dependencies_numeric}
\alias{plot_partial_dependencies_numeric}
\alias{pdp_class}
\title{Plot partial dependencies for the case of numeric dependenant variable
in the neural network}
\usage{
plot_partial_dependencies_numeric(predictor = "all", train,
  neural_network)

pdp_class(predictor, train, nn, class)
}
\arguments{
\item{predictor}{Vector of strings or numerics representing the independant variables
(as columns of the training set), default is "all"}

\item{train}{Training data set}

\item{neural_network}{Trained neural network}

\item{predictor}{Vector of strings representing the independant variables
(as columns of the training set), default is "all"}

\item{train}{Training data set}

\item{neural_network}{Trained neural network}
}
\value{
plot

plot
}
\description{
Plot the data

Plot the data
}
