% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{plot_single_numerical}
\alias{plot_single_numerical}
\title{Plots partial dependencies for given predictor with numerical dependent
variable.}
\usage{
plot_single_numerical(prepared_data, predictor, neural_net)
}
\description{
Plots partial dependencies for given predictor with numerical dependent
variable.
}
\keyword{internal}
