% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{plot_single_res}
\alias{plot_single_res}
\title{create plotting for single predictor}
\usage{
plot_single_res(neural_net, predictor, predictor_value, rep, units, kind,
  change_variables, change_values, class, probs, nrepetitions)
}
\description{
create plotting for single predictor
}
\keyword{internal}
