% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{plotting_input}
\alias{plotting_input}
\title{create input for plots}
\usage{
plotting_input(neural_net, predictor, predictor_value, rep, units, kind,
  change_variables, change_values, class, probs, nrepetitions)
}
\description{
create input for plots
}
\keyword{internal}
