% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{predict.NeuralNetwork}
\alias{predict.NeuralNetwork}
\title{Predict method for neural network.}
\usage{
\method{predict}{NeuralNetwork}(object, ...)
}
\arguments{
\item{object}{Fitted NeuralNetwork model}

\item{...}{further parameters for predict neuralnet, see:
\code{\link[neuralnet]{predict.nn}}}
}
\description{
Predict method for neural network.
}
