% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepares the data used for further plotting.}
\usage{
prepare_data(neural_net, predictor, probs = c(0.05, 0.95))
}
\arguments{
\item{neural_net}{Fitted NeuralNetwork, see: \code{\link{NeuralNetwork}}}

\item{predictor}{Predictor for which to prepare the plotting data (predictor
should be transformed with sym).}

\item{probs}{Vector of ower and upper bound probabilities for the confidence
interval.}
}
\value{
Data.frame with the prepared data.
}
\description{
\code{prepare_data} returns the prepared data used for further plotting.
}
\keyword{internal}
