% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{prepare_data_categoric}
\alias{prepare_data_categoric}
\title{Prepares the data with categorical dependent variable.}
\usage{
prepare_data_categoric(grid, predictor, neural_net, probs, nrepititions)
}
\description{
Prepares the data with categorical dependent variable.
}
\keyword{internal}
