% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{prepare_data_numeric}
\alias{prepare_data_numeric}
\title{Prepares the data with numerical dependent variable.}
\usage{
prepare_data_numeric(grid, predictor, neural_net, probs, nrepititions)
}
\description{
Prepares the data with numerical dependent variable.
}
\keyword{internal}
