% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R,
%   R/neural_net_representative_values.R
\name{prepare_data_repres}
\alias{prepare_data_repres}
\title{create plotting data for categorical and numerical variables}
\usage{
prepare_data_repres(neural_net, predictor, predictor_value, rep, units,
  kind, change_variables, change_values, class)

prepare_data_repres(neural_net, predictor, predictor_value, rep, units,
  kind, change_variables, change_values, class)
}
\description{
create plotting data for categorical and numerical variables

create plotting data for categorical variables
}
\keyword{internal}
