% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{prepare_numeric_data}
\alias{prepare_numeric_data}
\title{Preparation for the case of numeric dependenant variable in the
neural network}
\usage{
prepare_numeric_data(predictor, train, neural_network)
}
\arguments{
\item{predictor}{Vector of strings representing the independant variable
(as a column of the training set)}

\item{train}{Training data set}

\item{neural_network}{Trained neural network}
}
\value{
data.frame
}
\description{
Prepares the data for further plotting
}
