% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R
\name{rescale_total}
\alias{rescale_total}
\title{Rescale all selected variables}
\usage{
rescale_total(min, diff, predictor, data)
}
\description{
Rescale all selected variables
}
\keyword{internal}
