% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run the shiny app for visualizing the marginal effects of the NeuralNewtork}
\usage{
run_shiny_app()
}
\description{
Run the shiny app for visualizing the marginal effects of the NeuralNewtork
}
\examples{
\dontrun{
# Starts the shiny app
run_shiny_app()
}
}
