% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R,
%   R/neural_net_representative_values.R
\name{sequence}
\alias{sequence}
\title{Create sequence for prediction}
\usage{
sequence(x, rep, units)

sequence(x, rep, units)
}
\description{
Create sequence for prediction

create sequence for prediction
}
\keyword{internal}
