% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{summary.NeuralNetwork}
\alias{summary.NeuralNetwork}
\title{Summary method for neural network.}
\usage{
\method{summary}{NeuralNetwork}(object, ...)
}
\arguments{
\item{object}{Fitted NeuralNetwork model}

\item{...}{further parameters for summary method, see: \code{\link{summary}}}
}
\description{
Summary method for neural network.
}
