% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{warn_plot}
\alias{warn_plot}
\title{warns about length of predictors for plotting}
\usage{
warn_plot(type, predictor, predictor_value, len)
}
\description{
warns about length of predictors for plotting
}
\keyword{internal}
