% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R
\name{warning_message}
\alias{warning_message}
\title{issue warnings}
\usage{
warning_message(kind, predictor_value, data, predictor, change_variables,
  change_values)
}
\description{
issue warnings
}
\keyword{internal}
