% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R
\name{categ_prediction}
\alias{categ_prediction}
\title{Create prediction data for categorical variables}
\usage{
categ_prediction(neural_net, new_data, class, predictor, predictor_value,
  rep, units = NULL)
}
\description{
Create prediction data for categorical variables
}
\keyword{internal}
