% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{categorical_marg}
\alias{categorical_marg}
\title{create plotting data for categorical variables}
\usage{
categorical_marg(neural_net, data, predictor, predictor_value, rep, units,
  kind, change_variables, change_values, class, type)
}
\description{
create plotting data for categorical variables
}
\keyword{internal}
