% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{create_factor_specification}
\alias{create_factor_specification}
\title{Returns the data with added factor columns and changed formula}
\usage{
create_factor_specification(data, dependent, independent, f)
}
\description{
Returns the data with added factor columns and changed formula
}
\keyword{internal}
