% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_data_preparation.R
\name{descale}
\alias{descale}
\title{Returns the scale and the center of the predictor column.}
\usage{
descale(predictor, neural_net, grid)
}
\description{
Returns the scale and the center of the predictor column.
}
\keyword{internal}
