% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_class.R
\name{get_independent}
\alias{get_independent}
\title{Returrns the independent variables based on forumla.}
\usage{
get_independent(data, dependent_variable, specification)
}
\description{
Returrns the independent variables based on forumla.
}
\keyword{internal}
