% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{group_scale_multiple}
\alias{group_scale_multiple}
\title{scale grouped datasets, multiple datasets at once}
\usage{
group_scale_multiple(data, predictor, predictor_value)
}
\description{
scale grouped datasets, multiple datasets at once
}
\keyword{internal}
