% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{NeuralNetwork}
\alias{NeuralNetwork}
\title{Create a NeuralNetwork for further visualization}
\usage{
NeuralNetwork(f, data, layers, scale = FALSE, ...)
}
\arguments{
\item{f}{A formula representing the model that should be fitted. Handles
categorical, binary and numerical data.}

\item{data}{The data that should be used for training the neural network.}

\item{layers}{Vector representing the number of layers that should be used.}

\item{scale}{Boolean representing if the data should be scaled or not}

\item{...}{further parameters for neuralnet, see:
\code{\link[neuralnet]{neuralnet}}}
}
\value{
NeuralNetwork class containing the training data, the neural network
  and  the type of the dependent variable.
}
\description{
\code{NeuralNetwork} returns the trained neural network
}
\details{
This is a S3 class. It defines a neural network and the generic plot function
is overwritten to handle the NeuralNetwork class as input.
}
\examples{
# Example: Numeric
library(MASS)
neural_network <- NeuralNetwork(f = "medv ~ .", data = Boston,
                                layers = c(5, 3), scale = TRUE)

# Example: Categoric
library(datasets)
neural_network <- NeuralNetwork(f = "Species ~ .", data = iris,
                                layers = c(10, 10), rep = 5, err.fct = "ce",
                                linear.output = F, lifesign = "minimal",
                                stepmax = 1000000, threshold = 0.001)
}
