% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetworkVisualization.R
\docType{package}
\name{NeuralNetworkVisualization}
\alias{NeuralNetworkVisualization}
\alias{NeuralNetworkVisualization-package}
\title{NeuralNetworkVisualization: Create Partial Dependence Plots for Neural Networks.}
\description{
NeuralNetworkVisualization is a R Package for creating beautiful partial
  dependence plots for Neural Networks using ggplot2, plotly and shiny.
  It is possible to add confidence intervals created with a bootstrap
  procedure. There are three prefitted models that can be used for testing.
  You can acess them with the example_nn_model function.
}
