% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{numeric_marg}
\alias{numeric_marg}
\title{create plotting data for numeric}
\usage{
numeric_marg(neural_net, data, predictor, predictor_value, rep, units,
  kind, change_variables, change_values, class, type)
}
\description{
create plotting data for numeric
}
\keyword{internal}
