% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R,
%   R/neural_net_visualization.R
\name{plot_multiple}
\alias{plot_multiple}
\title{create plotting for multiple predictors}
\usage{
plot_multiple(neural_net, predictors, probs, nrepetitions)

plot_multiple(neural_net, predictors, probs, nrepetitions)
}
\description{
create plotting for multiple predictors

Plots partial dependencies for multiple predictors.
}
\keyword{internal}
