% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{plot_multiple}
\alias{plot_multiple}
\title{Plots partial dependencies for multiple predictors.}
\usage{
plot_multiple(neural_net, predictors, probs, nrepetitions)
}
\description{
Plots partial dependencies for multiple predictors.
}
\keyword{internal}
