% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_visualization.R
\name{plot_single}
\alias{plot_single}
\title{Plots partial dependencies for single given predictor.}
\usage{
plot_single(neural_net, predictor, probs, nrepetitions)
}
\description{
Plots partial dependencies for single given predictor.
}
\keyword{internal}
