% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative.R
\name{prepare_data_repres}
\alias{prepare_data_repres}
\title{create plotting data for categorical variables}
\usage{
prepare_data_repres(neural_net, predictor, predictor_value, rep = 100,
  units = 1, kind = "mean", change_variables = NULL,
  change_values = NULL, class = NULL)
}
\description{
create plotting data for categorical variables
}
\keyword{internal}
