% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{scale_change}
\alias{scale_change}
\title{scale change_variables}
\usage{
scale_change(data, change_variables, change_values)
}
\description{
scale change_variables
}
\keyword{internal}
