% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_net_representative_values.R
\name{warning_message}
\alias{warning_message}
\title{issue warnings}
\usage{
warning_message(neural_net, kind, predictor_value, predictor,
  change_variables, change_values, type, class)
}
\description{
issue warnings
}
\keyword{internal}
